// src/app/models/customer.model.ts
export interface Customer {
    _id: string;
    name: string;
    phone1: string;
    phone2: string;
    email1: string;
    email2: string;
    address: string;
    city: string;
    state: string;
    country: string;
    zipcode: string;
    gstn: string;
    pan: string;
    type: string;
}